proyecto-iot/requirements.txt
Flask
Flask-CORS
paho-mqtt
